+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+Unreleased
+----------
+- Renamed `data::v2::bars::BarReq` to `BarsReq`
+  - Deprecated `data::v2::bars::BarReq`
+
+
 0.22.4
 ------
 - Adjusted `Subscribable` trait to make all created futures implement